requirements.txt
pandas==1.1.4
pymongo==3.11.1
Flask
python-dotenv
